refactoring ltr passage and doc
