+++ docs/experiments-ltr-msmarco-passage-reranking.md
@@ -52,12 +52,16 @@ tar -xzvf runs/msmarco-passage-ltr-mrr-v1.tar.gz -C runs
 Next we can run our inference script to get our reranking result.
 
 ```bash
-python -m pyserini.ltr.search_msmarco_passage \
+python scripts/ltr_msmarco-passage/ltr_inference.py \
   --input runs/run.msmarco-passage.bm25tuned.txt \
   --input-format tsv \
   --model runs/msmarco-passage-ltr-mrr-v1 \
   --index ~/.cache/pyserini/indexes/index-msmarco-passage-ltr-20210519-e25e33f.a5de642c268ac1ed5892c069bdc29ae3 \
+  --data passage \
+  --ibm-model collections/msmarco-ltr-passage/ibm_model/ \
+  --queries collections/msmarco-ltr-passage \
   --output runs/run.ltr.msmarco-passage.tsv 
+
 ```
 
 Here, our model is trained to maximize MRR@10.
